+++ R/nutpieR-package.R
@@ -1,2 +1,3 @@
 #' @keywords internal
+#' @useDynLib nutpieR, .registration = TRUE
 "_PACKAGE"